requirements.txt
asgiref
autopep8
Django==2.1.*
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
psycopg2
django-import-export
forex_python
pandas
scipy
sklearn
